+++ scripts/installPackages.sh
@@ -4,15 +4,18 @@
 if [ $CI ] || [ $TASK_DEFINITION ]
 then
   echo "Running in a CI environment"
-  YARN_ARGS="--frozen-lockfile"
+  YARN_INSTALL_ARGS="--frozen-lockfile"
 else
   echo "Running in a local dev environment"
-  YARN_ARGS=""
+  YARN_INSTALL_ARGS=""
 fi
 
-# Creates symlinks between internal dependencies. Also runs the 'prepare' script
-npx lerna bootstrap --force-local -- $YARN_ARGS
+# Install dependencies and run the 'prepare' script
+# Running yarn in each package directory to create a unique yarn.lock file
+npx lerna exec --concurrency=1 -- yarn install $YARN_INSTALL_ARGS
+npx lerna run prepare
 
-# Calling 'yarn install' within a package directory may overwrite a symlink with a hard copy.
+# Calling 'yarn install' within a package directory may install a hard copy from npm repository
+# Overwrite a hard copy with a symlink. This ensures we are always running and testing against the local version
 npx lerna link --force-local