+++ packages/configuration-validation/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.4.1
+
+### Bug Fixes
+
+- [`a2a7b3e`](https://github.com/okta/okta-oidc-js/commit/a2a7b3e695d40e29d473be89e90340fbf5c4c56b) - Normalize config format for the properties `responseType` and `scopes`, used in get token flows.
+
+# 0.4.0
+
+### Features
+
+ - [`654550`](https://github.com/okta/okta-oidc-js/commit/6545506921cbe6e8f15076e45e908f285a6e2f1e) - All configuration options are now accepted. See [Configuration Reference](https://github.com/okta/okta-auth-js#configuration-reference). Camel-case (clientId) is now the preferred syntax for all Okta OIDC libraries. Underscore syntax (client_id) may be deprecated in a future release.
+
 # 0.3.0
 
 ### Features